XD-1846 Simplify DeploymentVerifier
- With the Stream/Job Deployment State changes, the `DeploymentVerifier`
has been simplified to use the DeploymentUnitStatus.State to verify if
the Stream/Job is successfully deployed/undeployed.
    - Add `getDeploymentStatus` in DeploymentPathProvider and remove other
deployment path specific method contracts.
    - Have `Stream/JobPathProvider`s implement this method and use Stream/JobRepository's
`getDeploymentStatus` (via DeploymentStatusRepository) to return DeploymentUnitStatus
    - Since the StateCalcualtor already determines the number of modules that need to be
deployed for a given deployment unit, the verification of `DeploymentUnitStatus.State`
can be sufficient to assert if the deployment unit is successfully deployed/undeployed
with the expected number of modules count etc.,
    - Update `DeploymentVerifier`'s deployment verifier methods waitForDeploy/waitForUndeploy