XD-1944: Error deploying stream when no matching containers
When a container arrives each stream is examined to determine which
modules have been deployed to it. If no modules have been deployed
we get a NoNodeException for the path
/xd/deployments/streams/<name>/modules. The exception is now trapped
and ignored as it simply indicates that no modules have yet been
deployed for the stream.

Handle case when deploying a stream without containers

Wrap a try/catch NoContainerException around every attempt to
match a container - otherwise some of the module requests
will not be written and the stream state will not be correctly
calculated.