Appends file: to XD_CONFIG_LOCATION
When user sets XD_CONFIG_LOCATION modules would fail because they could not find the config directory.
This was because the first / was missing.  To resolve this "file:" is appended to the XD_CONFIG_LOCATION.

To reproduce:
1. export XD_CONFIG_LOCATION=<config directory
2. deploy the following: 'stream create foo --definition "time|hdfs" --deploy .

The hdfs will not deploy properly because it can not find hadoop.properties.

This should resolve the issue.