XD-1958 Fix Partitioned Batch Jobs
JIRA: https://jira.spring.io/browse/XD-1958

With the `RabbitMessageBus`, the `replyChanel` header was
not propagated and so the aggregator failed to deliver
the step execution results to the partition handler.

Use a `HeaderChannelRegistry` to propagate a key
referencing the `replyChannel`.

With the `RedisMessagebus` there was no provision to
propagate *any* headers needed for paritioning
(`correlationId`, `sequenceSize` etc).

Add a list of standard headers to propagate and a
facility to add additional headers (at the bus level).

Add tests to assert propagation; also tested manually
with the file->jdbc job with 3 partitions (both Rabbit
and Redis).