XD-1983 NodeExistException on ModuleMetadata write
- Ignore ZK `NodeExistsException` on writing module metadata as this
is likely to happen when the container disconnects and reconnects
but the admin leader was not available to handle the container
departure event.
  - In this case, upon container disconnect the modules on that
container get undeployed and when the container reconnects
the `DeploymentListener` makes sure to deploy the same modules
except the fact that the ZK paths for the deployment and module
metadata already exists.

2nd commit message:

Fix code review comment

- Make sure to update the status even when `NodeExistsException`
is thrown while writing module metadata path.