Added synchronization for container attributes
Changed ContainerAttributes internal map to ConcurrentHashMap
to take into account the possibility of multiple threads
reading & writing attributes.

When the admin port for a container is extracted from
an application context event, the port number will
be updated in the container's ephemeral node if the
node already exists. If the node does not exist, it
is assumed that the container is in the middle of
the registration process on another thread and therefore
the container attributes (including the management port)
will eventually be written to the container node.

The update to container attributes, check for node, and
update of node is now an atomic operation. Additionally,
the serialization of container attributes and subsequent
creation of the container node is also an atomic operation.
This guarantees that operations requiring read/write of
container attributes followed by a write to ZooKeeper
will happen on one thread at a time.

Improved assert message in functional test.