XD-2169 - Create standalone list of deps for project
Switch from platform-managed dependency versions to a local file
compatible with the `versionManagement` mechanism of the Spring Boot
Gradle Plugin (initially, the file is a clone of the versions list
found in Spring IO 1.0.1.RELEASE).

This is necessary in order to allow the project's dependency list to
evolve independently from Spring IO (e.g. in order to upgrade Spring Boot).