XD-2170: NoNodeException for job creation
When an admin process is elected leader/supervisor, it will query
the state of all streams and jobs. It does this by iterating all
streams under /xd/deployments/streams/[stream-name] and jobs under
/xd/deployments/jobs/[job-name]. Under each path, it is expecting a
modules node which contains all of the modules deployed for the stream or job.

The issue is that if *no* modules were deployed for the stream/job,
that node will not exist because it is created lazily by the container.
This is unlikely in the case of streams since a stream contains at
least two modules, but is much more likely with jobs since they
only contain one module.

The fix is to ensure that the
/xd/deployments/streams/[stream-name]/modules path (and the same for
jobs) is created upfront by the supervisor prior to creating the
module deployment requests. This will allow the supervisor to see
that no modules for the stream/job have been deployed and thus
take appropriate action.