XD-2286: Simple OOTB job for testing
The "timestampfile" job will append the current date/time to a file
when launched. This job can be used to validate a Spring XD
installation and to verify that jobs can be deployed. This serves a
similar purpose to "time | log" for streams.

This job supports the following options:

fileName
  Name of file to write timestamps to. Excludes parent directory and
  file extension. Defaults to the job name.

directory
  Name of directory to write timestamp file to. Defaults to
  "/tmp/xd/output/".

fileExtension
  Extension for file to write timestamps to. Defaults to "txt".

format
  How to render the current time, using SimpleDateFormat. Defaults to
  "yyyy-MM-dd HH:mm:ss"

restartable
  true if the job may be restarted.  false by default