Removed separate deployer thread
In commit a272d51 a separate thread for stream deployments was
added in order to avoid slowing down the Curator thread responsible
for cache updates. However this may lead to a race condition where
the Curator thread (from handling a container arrival event)
may deploy a module at the same time as the dedicated initial deployment
thread.

The fix is to remove the separate initial deployment thread and
force all deployments, regardless of source, into a single thread.