XD-2423 Defer WireTap Until Tap Channel Bound
JIRA: https://jira.spring.io/browse/XD-2423

Previously, the `WireTap` was added to the module `output` channel
before the channel was bound to the bus,

There is a race condition where the main stream can start sending
to the tap before it is bound to the bus,  causing 'Dispatcher has no subscribers`
errors. In 1.0.x this would prevent the message from being sent to the
main stream. In 1.1.x, taps run on a different thread so it's just log
noise.

Defer adding the `WireTap` to the tapped module's `output` channel
until it has been bound to the bus as a pub/sub producer.