http source now accepts binary content larger that a few KB
The problem was the assumption that the ChannerBuffer is always backed
by an array - which is not the case if you send e.g. binary data that
is larger than around 10KB. The solution is to check the availability
of a backing byte[] with the hasArray() method.

Note that after this fix the source is still not able to accept
messages > ~1MB in size. This affects all messages, binary and others.
I guess that this must be somehow fixed in the netty setup ? I’ve just
not investigated this since this fix solves my current problem.