XD-2590: Handle Large Messages in HTTP Source
JIRA: https://jira.spring.io/browse/XD-2590

Based on original work by Peter Rietzler:

http source now accepts binary content larger that a few KB

The problem was the assumption that the ChannerBuffer is always backed
by an array - which is not the case if you send e.g. binary data that
is larger than around 10KB. The solution is to check the availability
of a backing byte[] with the hasArray() method.

- move the decoding work to a `MessageConverter`
- allow the message converter to be configured
- add a setter for the max content length; add error handling when the content length is exceeded
  - previously, the client hung waiting for a reply.
  - close the channel in the catch all `exceptionCaught` method and log the error.
- add properties to the http source (maxContentLength, messageConverterClass)