XD-2606: Fix Track History with Redis/Kafka
Add `xdHistory` to the standard list of headers passed by the bus,

Fix `EmbeddedHeaderMessageConverter` to handle non-string headers and
increase the size of the header value to `Integer.MAX_VALUE` (from 255).

Store the header values as JSON; eliminates the need to interpret
specific headers, allows transport of the `xdHistory` collection.

The converter will continue to decode "old" messages that might still
be in the bus before upgrading.