XD-2674 Avoid escaping topic name twice
Due to a bug in KafkaMessageBus.escapeTopicName(), "_" is re-converted on already escaped topic names (even if it is a valid character). This causes the initialization of the topic states to read the wrong initial offsets, subsequently causing the test to fail. This behaviour is observable only on already cleaned (pruned) topics.