XD-2486 Remove rehydrating EC when launching a job.
When a custom job is launched, the classpath for the custom job is not
available to the JobLauncher.  Because of this, if there are custom
classes that have been serialzied in the ExecutionContext, a ClassNotFoundException is
thrown.  Since the EC isn't needed to launch a job, this fix removes the
rehydration of the EC by the launcher.