XD-2814: Durable Pub/Sub with Rabbit Bus
JIRA: https://jira.spring.io/browse/XD-2814

The `RabbitMessageBus` now supports durable queues for pub/sub named channels
(`tap:` and `topic:`). Such queues are eligible for automatic DLQ creation
and binding.

This change also means that multiple instances of the first module in a stream
when the input channel is a named pub/sub are now competing consumers for
messages in that channel.

Instead of naming the queue with a UUID, the stream name now becomes part of the
name:

    xdbus.<stream>.tap:stream:foo.0

and

    xdbus.<stream>.tap:stream:foo.0.dlq

Previously, each instance received all messages which was incorrect (XD-2115).

XD-2814: Initial Commit

Use Capability

Tests and DLQ Setup

Docs

Fix Exclusive Bit When Not Durable

Now that we support competing consumers on taps, the queue for
a non-durable subscription cannot be exclusive (but is still auto-delete).