XD-2911 - Improved performance of Tuple Creation
This commit refactors how the conversion service used within a Tuple is
created to a more performant option.  Prior to this commit, for each
instance of TupleBuilder created (and under typical usage, each Tuple)
an instance of DefaultTupleConversionService was created.  This updated
version changes the default to be a static instance and changes the
recommended approach for customization to be a
ConfigurableConversionService be configured as a singleton within your
module's context and then used across all Tuple instances.  This
provides better performance in the creation of Tuples.