XD-3117: Add Logging to ZooKeeperContainerRepository
If the PathChildrenCache for containers is closed due
to a PathChildrenCacheEvent indicating a ZooKeeper
disconnect, the null assert will fail. This change
is for the event to log a message so that it becomes
clear that the assert failed because of a closed or
suspended connection.

Log connection interruptions as WARN